feat: add time formatting utility functions and refactor attendance components to use them
