+++ apps/frontend/src/app/(admin)/check-clock/_tabsContent/CheckClockOverviewTab.tsx
@@ -12,6 +12,7 @@ import { PageSizeComponent } from "@/components/pageSize";
 import { AttendanceDetailSheet } from "../_components/AttendanceDetailSheet";
 import { LeaveRequestDetailSheet } from "../_components/LeaveRequestDetailSheet";
 import { AddAttendanceDialog } from "../_components/AddAttendanceDialog";
+import { formatWorkHours, formatTime } from "@/utils/time";
 import * as React from "react";
 import {
 	ColumnDef,
@@ -168,25 +169,7 @@ export default function CheckClockOverviewTab() {
 				accessorKey: "clock_in",
 				cell: ({ row }) => {
 					if (row.original.type === "leave_request") return "-";
-
-					const clockIn = row.original.clock_in;
-					if (!clockIn) return "-";
-
-					// Handle HH:MM:SS format
-					if (clockIn.match(/^\d{2}:\d{2}:\d{2}$/)) {
-						return clockIn.substring(0, 5);
-					}
-
-					try {
-						const time = new Date(clockIn);
-						return time.toLocaleTimeString("en-US", {
-							hour: "2-digit",
-							minute: "2-digit",
-							hour12: false,
-						});
-					} catch {
-						return clockIn.substring(0, 5);
-					}
+					return formatTime(row.original.clock_in);
 				},
 			},
 			{
@@ -194,25 +177,7 @@ export default function CheckClockOverviewTab() {
 				accessorKey: "clock_out",
 				cell: ({ row }) => {
 					if (row.original.type === "leave_request") return "-";
-
-					const clockOut = row.original.clock_out;
-					if (!clockOut) return "-";
-
-					// Handle HH:MM:SS format
-					if (clockOut.match(/^\d{2}:\d{2}:\d{2}$/)) {
-						return clockOut.substring(0, 5);
-					}
-
-					try {
-						const time = new Date(clockOut);
-						return time.toLocaleTimeString("en-US", {
-							hour: "2-digit",
-							minute: "2-digit",
-							hour12: false,
-						});
-					} catch {
-						return clockOut.substring(0, 5);
-					}
+					return formatTime(row.original.clock_out);
 				},
 			},
 			{
@@ -220,13 +185,7 @@ export default function CheckClockOverviewTab() {
 				accessorKey: "work_hours",
 				cell: ({ row }) => {
 					if (row.original.type === "leave_request") return "-";
-
-					const hours = row.original.work_hours;
-					if (!hours) return "-";
-
-					const wholeHours = Math.floor(hours);
-					const minutes = Math.round((hours - wholeHours) * 60);
-					return `${wholeHours}h ${minutes}m`;
+					return formatWorkHours(row.original.work_hours);
 				},
 			},
 			{
